actorwithbody splitted in favour of new particle type
Former-commit-id: 121bd069d0a9eeef60f5ecb085a11a93c4b4a84d
Former-commit-id: 539b4b6916e808c01298190cf347e928f61fe62e

+++ src/net/torvald/terrarum/mapgenerator/NoiseFilter.kt
@@ -4,5 +4,5 @@ package net.torvald.terrarum.mapgenerator
  * Created by minjaesong on 16-03-31.
  */
 interface NoiseFilter {
-    fun getGrad(func_argX: Int, start: Float, end: Float): Float
+    fun getGrad(func_argX: Int, start: Double, end: Double): Double
 }
