+++ src/net/torvald/terrarum/mapgenerator/NoiseFilterUniform.kt
@@ -4,7 +4,7 @@ package net.torvald.terrarum.mapgenerator
  * Created by minjaesong on 16-03-31.
  */
 object NoiseFilterUniform : NoiseFilter {
-    override fun getGrad(func_argX: Int, start: Float, end: Float): Float {
-        return 1f
+    override fun getGrad(func_argX: Int, start: Double, end: Double): Double {
+        return 1.0
     }
 }
